Modernize project
