--- .hound.yml
@@ -1,2 +1,0 @@
-swift:
-  config_file: .swiftlint.yml
