--- .swift-version
@@ -1,1 +1,0 @@
-4.0
