--- Example/Podfile
@@ -1,11 +1,0 @@
-use_frameworks!
-
-target 'Swen_Example' do
-  pod 'Swen', :path => '../'
-
-  target 'Swen_Tests' do
-    inherit! :search_paths
-
-    
-  end
-end
